requirements.txt
# local package
-e .

# external requirements
click~=8.0.3
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
black
pandas~=1.3.5
numpy~=1.21.2
transformers~=4.15.0
datasets~=1.17.0
scipy~=1.7.3
scikit-learn~=1.0.2
torch
mlflow
matplotlib
setuptools
